instagram/requirements.txt
selenium
webdriver-manager
discord.py
